api: routes에 들어있는 소켓 관련 코드를 분리
