--- api/core/service/group/index.js
@@ -1,10 +1,0 @@
-const findGroups = () => {
-  return [
-    { id: 1, name: '개발실' },
-    { id: 2, name: '플랫폼 개발팀', groupId: 1 },
-  ];
-};
-
-module.epoxrts = {
-  findGroups,
-};
